chore: move `get_orchestrator_client` from `substrapp.orchestrator` to `orchestrator`
